module2-task1: * ts setting
