+++ package.json
@@ -13,6 +13,7 @@
     "@typescript-eslint/eslint-plugin": "3.7.1",
     "eslint": "6.1.0",
     "eslint-config-airbnb-typescript": "9.0.0",
+    "eslint-import-resolver-alias": "1.1.2",
     "eslint-plugin-import": "2.22.0",
     "ts-loader": "8.0.1",
     "typescript": "3.9.7",
