+++ src/view/load-more-button.ts
@@ -1,0 +1,5 @@
+const createLoadMoreButtonTemplate = () => `<button class="load-more" type="button">
+    load more
+  </button>`;
+
+export { createLoadMoreButtonTemplate };
